+++ src/App.tsx
@@ -11,6 +11,7 @@ import { CustomerDetailPage } from '@/routes/CustomerDetailPage'
 import { ClinicsPage } from '@/routes/ClinicsPage'
 import { ClinicDetailPage } from '@/routes/ClinicDetailPage'
 import { PrimPage } from '@/routes/PrimPage'
+import { SamplesPage } from '@/routes/SamplesPage'
 import { CariHesapPage } from '@/routes/CariHesapPage'
 import { CariHesapListPage } from '@/routes/CariHesapListPage'
 import { BudgetYearPage } from '@/routes/BudgetYearPage'
@@ -59,6 +60,7 @@ function App() {
             <Route path="klinikler" element={<ClinicsPage />} />
             <Route path="klinikler/:id" element={<ClinicDetailPage />} />
             <Route path="prim" element={<PrimPage />} />
+            <Route path="numuneler" element={<SamplesPage />} />
             <Route path="cari-hesap" element={<CariHesapListPage />} />
             <Route path="cari-hesap/:id" element={<CariHesapPage />} />
             <Route path="stok" element={<StockPage />} />